Make keepright loader into a real module
