+++ index.html
@@ -9,7 +9,6 @@
     <script src='http://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.1/jquery.min.js'></script>
     <script src='http://api.tiles.mapbox.com/mapbox.js/plugins/leaflet-osm/v0.1.0/leaflet-osm.js'></script>
     <link rel="stylesheet" type="text/css" href="./css/site.css">
-    <link rel="stylesheet" type="text/css" href="./css/colors.min.css">
 </head>
 <body id='tofix' class='fill-grey'>
     <div id="error-message" class="center note error">&nbsp;<span></span>&nbsp;</div>
@@ -18,13 +17,6 @@
     <div id='main' class='unlimiter contain'>
         <div id='settings' class='animate modal modal-content'></div>
         <div id='map' class='fill-black'></div>
-        <div id='bottom-bar' class='col12 pin-bottom'>
-            <div id='actions' class='col6 margin3 fill-dark'>
-                <h3 id='edit' class='col4 pad2 center'>edit</h3>
-                <h3 id='skip' class='col4 pad2 center'>skip</h3>
-                <h3 id='fixed' class='col4 pad2 center'>fixed</h3>
-            </div>
-        </div>
     </div>
     <script src='js/bundle.js'></script>
 </body>
