+++ templates/editbar.html
@@ -1,0 +1,7 @@
+<div id='bottom-bar' class='col12 pin-bottom'>
+    <div id='actions' class='col6 margin3 fill-dark'>
+        <h3 id='edit' class='col4 pad2 center'>edit</h3>
+        <h3 id='skip' class='col4 pad2 center'>skip</h3>
+        <h3 id='fixed' class='col4 pad2 center'>fixed</h3>
+    </div>
+</div>
